pyproject.toml
[project]
name = "azure-custom-evaluators"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.12"
dependencies = [
    "azure-ai-evaluation",
    "promptflow",
    "ipykernel>=6.29.5",
    "jedi<0.19",
    "python-dotenv>=1.1.1",    
]
